Add files via upload
Branch for combat. It should work, but by canvas is fucking up so I can't test it.

Just run the testArena file, it has a demo that creates a few hives that we can run.

Right now, it is not run at all in the main code, that can only happen once the list_hives and list_ants are properly implemented.